# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Job Scout Pro AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
    put:
      operationId: upsertProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # ── Resume ───────────────────────────────────────────────────────────────────
  /resume:
    get:
      operationId: getResume
      tags: [resume]
      summary: Get the current user's resume
      responses:
        "200":
          description: Resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found
    post:
      operationId: uploadResume
      tags: [resume]
      summary: Upload and parse resume text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "200":
          description: Parsed resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

  # ── Jobs ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, userId, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        fullName:
          type: ["string", "null"]
        preferredRoles:
          type: array
          items:
            type: string
        preferredLocations:
          type: array
          items:
            type: string
        expectedSalaryMin:
          type: ["integer", "null"]
        expectedSalaryMax:
          type: ["integer", "null"]
        workMode:
          type: ["string", "null"]
          enum: ["remote", "hybrid", "onsite", null]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ProfileInput:
      type: object
      properties:
        fullName:
          type: string
        preferredRoles:
          type: array
          items:
            type: string
        preferredLocations:
          type: array
          items:
            type: string
        expectedSalaryMin:
          type: integer
        expectedSalaryMax:
          type: integer
        workMode:
          type: string
          enum: ["remote", "hybrid", "onsite"]

    # ── Resume ──────────────────────────────────────────────────────────────────
    Resume:
      type: object
      required: [id, userId, rawText, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        rawText:
          type: string
        skills:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        projects:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ResumeInput:
      type: object
      required: [rawText]
      properties:
        rawText:
          type: string
          minLength: 50

    # ── Job ─────────────────────────────────────────────────────────────────────
...