# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Job Scout Pro AI API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileWorkMode } from './profileWorkMode';

export interface Profile {
  id: number;
  userId: string;
  /** @nullable */
  fullName?: string | null;
  preferredRoles?: string[];
  preferredLocations?: string[];
  /** @nullable */
  expectedSalaryMin?: number | null;
  /** @nullable */
  expectedSalaryMax?: number | null;
  /** @nullable */
  workMode?: ProfileWorkMode;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}