# lib/db/src/schema/resumes.ts
import { pgTable, serial, text, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const resumesTable = pgTable("resumes", {
  id: serial("id").primaryKey(),
  userId: text("user_id").notNull().unique(),
  rawText: text("raw_text").notNull(),
  skills: text("skills").array().notNull().default([]),
  experience: text("experience").array().notNull().default([]),
  education: text("education").array().notNull().default([]),
  certifications: text("certifications").array().notNull().default([]),
  projects: text("projects").array().notNull().default([]),
  createdAt: timestamp("created_at").notNull().defaultNow(),
  updatedAt: timestamp("updated_at"),
});

export const insertResumeSchema = createInsertSchema(resumesTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertResume = z.infer<typeof insertResumeSchema>;
export type Resume = typeof resumesTable.$inferSelect;